v32 of timeit.py

+++ timeit.py
@@ -1,8 +1,8 @@
 import time
 
 def calculate_time(time):
-	time_start = float(time())
-	time_end = float(time())
+	time_start = int(time())
+	time_end = int(time())
 	total_time = time_end - time_start
 	print ('Total time', total_time)
